# service/src/routes/orders.js
'use strict';

const crypto = require('crypto');
const express = require('express');
const router = express.Router();

const orderStore = require('../store/order-store');
const pickupStore = require('../store/pickup-store');
const cancellationStore = require('../store/cancellation-store');
const { toOrderRepresentation } = require('../representation/order');
const { toCancellationRepresentation } = require('../representation/cancellation');
const {
  validateOrderId,
  validateCreateOrder,
  validateListParams,
} = require('../schemas/order-schema');
const {
  sendProblem,
  badRequest,
  notFound,
  orderNotCancellable,
  unprocessable,
  preconditionFailed,
  validationExtras,
} = require('../problem');
const { requireScope } = require('../auth/require-scope');
const {
  mayReadOrder,
  mayCancelOrder,
  mayCreateOrder,
  mayClaimOrder,
} = require('../auth/ownership');
const {
  orderVersion,
  collectionVersion,
  matchesWeak,
  matchesStrong,
} = require('../conditional');
const {
  checkIdempotency,
  claimIdempotency,
  saveIdempotencyRecord,
} = require('../idempotency');

function generateId(prefix) {
  const ts = Date.now().toString(36).toUpperCase();
  const rand = crypto.randomBytes(4).toString('hex').toUpperCase().slice(0, 6);
  return `${prefix}${ts}${rand}`;
}

function sendOrderNotFound(res) {
  // Identical for absent and not-owned, including the instance: a differing
  // instance path would itself leak which identifiers exist.
  return sendProblem(res, notFound('/v1/orders/{orderId}'));
}

router.get('/:orderId', requireScope('orders:read'), (req, res) => {
  const { orderId } = req.params;
  const err = validateOrderId(orderId);
  if (err) return sendProblem(res, badRequest(err, req.originalUrl));

  const row = orderStore.getById(orderId);
  if (!row) return sendOrderNotFound(res);

  const assignment = pickupStore.getByOrderId(orderId);
  if (!mayReadOrder(req.principal, row, assignment)) {
    return sendOrderNotFound(res);
  }

  const tag = orderVersion(row);
  res.set('ETag', tag);

  // Evaluated explicitly rather than left to Express's `req.fresh`.
  // `fetch()` injects `Cache-Control: no-cache` whenever it sends
  // `If-None-Match`, and `fresh` reports a `no-cache` request as stale by
  // design, so relying on it means the 304 never fires from a browser or from
  // any test written with `fetch`. See docs/decisions for the measurement.
  if (matchesWeak(req.headers['if-none-match'], tag)) {
    return res.status(304).end();
  }

  res.status(200).json(toOrderRepresentation(row));
});

router.get('/', requireScope('orders:read'), (req, res) => {
  const { status, limit, cursor } = req.query;

  const errors = validateListParams({ status, limit });
  if (errors) {
    return sendProblem(res, badRequest(errors.join('; '), req.originalUrl));
  }

  const parsedLimit = limit ? parseInt(limit, 10) : 20;
  const result = orderStore.listForPrincipal(req.principal, {
    status,
    limit: parsedLimit,
    cursor,
  });

  if (result === null) {
    return sendProblem(res, badRequest('Invalid cursor value.', req.originalUrl));
  }

  if (result.nextCursor) {
    res.set('X-Next-Cursor', result.nextCursor);
  }

  const items = result.rows.map(toOrderRepresentation);
  const tag = collectionVersion(items);
  res.set('ETag', tag);

  // A polled collection: when nothing changed the caller gets 304 with no
  // body, which is the point of conditional reads (P5 §A.7).
  if (matchesWeak(req.headers['if-none-match'], tag)) {
    return res.status(304).end();
  }

  res.status(200).json(items);
});

router.post('/', requireScope('orders:write'), (req, res) => {
  const idem = checkIdempotency(req);
  if (idem.error) {
    if (idem.retryAfter) res.set('Retry-After', String(idem.retryAfter));
    return sendProblem(res, idem.error);
  }
  if (idem.replay) {
    const cached = JSON.parse(idem.body);
    if (cached.id) res.set('Location', `/v1/orders/${cached.id}`);
    return res.status(idem.status).json(cached);
  }

  const validation = validateCreateOrder(req.body);
  if (validation) {
    const extras = validationExtras(validation.errors);
    const problem =
      validation.status === 422
        ? unprocessable(validation.errors.join('; '), req.originalUrl, extras)
        : badRequest(validation.errors.join('; '), req.originalUrl, extras);
    return sendProblem(res, problem);
  }

  if (!mayCreateOrder(req.principal, req.body.customerId)) {
    // Same 404 shape as an order that is not yours: a customer may only create
    // orders for their own domain identity, and the refusal must not reveal
    // whether the named customer exists.
    return sendProblem(res, notFound('/v1/orders/{orderId}'));
  }

  const claim = claimIdempotency(idem.key, idem.hash, req.originalUrl);
  if (claim) {
    if (claim.retryAfter) res.set('Retry-After', String(claim.retryAfter));
    if (claim.replay) return res.status(claim.status).json(JSON.parse(claim.body));
    return sendProblem(res, claim.error);
  }

  const now = new Date().toISOString();
  const orderId = generateId('ord_');
  const order = {
    id: orderId,
    customer_id: req.body.customerId,
    service_type: req.body.serviceType,
    weight_kg: req.body.weightKg,
    pickup_address: req.body.pickupAddress,
    status: 'pending_pickup',
    outlet_id: null,
    created_at: now,
    updated_at: now,
  };

  orderStore.insert(order);

  const representation = toOrderRepresentation(order);
  saveIdempotencyRecord(idem.key, idem.hash, 201, JSON.stringify(representation));

  res
    .status(201)
    .set('Location', `/v1/orders/${orderId}`)
    .json(representation);
});

const CANCELLABLE = ['pending_pickup', 'ready_for_pickup', 'confirmed'];

router.post('/:orderId/cancellation', requireScope('orders:write'), (req, res) => {
  const { orderId } = req.params;

  const idErr = validateOrderId(orderId);
  if (idErr) return sendProblem(res, badRequest(idErr, req.originalUrl));

  const idem = checkIdempotency(req);
  if (idem.error) {
    if (idem.retryAfter) res.set('Retry-After', String(idem.retryAfter));
    return sendProblem(res, idem.error);
  }
  if (idem.replay) {
    return res.status(idem.status).json(JSON.parse(idem.body));
  }

  const order = orderStore.getById(orderId);
  if (!order || !mayCancelOrder(req.principal, order)) {
    return sendProblem(res, notFound('/v1/orders/{orderId}/cancellation'));
  }

  // Precondition before the state check, as RFC 9110 evaluates it before the
  // method is performed: "the thing you read is no longer what is there" is
  // more precise than "this order is in the wrong state", and it is the
  // condition the client can actually act on.
  //
  // Optional but honoured. Existing clients that never send If-Match keep
  // working unchanged; a client that does send it can no longer overwrite a
  // change it has not seen (P5 §A.8.1).
  const ifMatch = req.headers['if-match'];
  if (ifMatch && !matchesStrong(ifMatch, orderVersion(order))) {
    return sendProblem(
      res,
      preconditionFailed(
        'Order ini sudah berubah sejak terakhir dibaca. Muat ulang untuk melihat keadaan terbaru.',
        req.originalUrl,
      ),
    );
  }

  if (!CANCELLABLE.includes(order.status)) {
    return sendProblem(
      res,
      orderNotCancellable(
        `Order cannot be cancelled because it is in '${order.status}' status.`,
        req.originalUrl,
        order.status,
        CANCELLABLE,
      ),
    );
  }

  const claim = claimIdempotency(idem.key, idem.hash, req.originalUrl);
  if (claim) {
    if (claim.retryAfter) res.set('Retry-After', String(claim.retryAfter));
    if (claim.replay) return res.status(claim.status).json(JSON.parse(claim.body));
    return sendProblem(res, claim.error);
  }

  const now = new Date().toISOString();
  orderStore.updateStatus(orderId, 'cancelled', now);

  const cancellation = {
    id: generateId('can_'),
    order_id: orderId,
    reason: 'Customer requested cancellation',
    status: 'completed',
    created_at: now,
  };
  cancellationStore.insert(cancellation);

  const representation = toCancellationRepresentation(cancellation);
  saveIdempotencyRecord(idem.key, idem.hash, 200, JSON.stringify(representation));

  res.status(200).json(representation);
});

// Staff-only operation (scope `orders:fulfil`). A customer token is refused at
// Layer 2 with 403 before the database is touched. Staff claim an unbound
// order, or keep working on an order already bound to their outlet; an order
// bound to another outlet is answered with the same 404 as an absent order.
router.post('/:orderId/fulfilment', requireScope('orders:fulfil'), (req, res) => {
  const { orderId } = req.params;

  const err = validateOrderId(orderId);
  if (err) return sendProblem(res, badRequest(err, req.originalUrl));

  const order = orderStore.getById(orderId);
  if (!order) return sendOrderNotFound(res);

  // Ownership is decided before any mutation is attempted.
  if (!mayClaimOrder(req.principal, order)) {
    return sendOrderNotFound(res);
  }

  // Conditional write: without a precondition the server has no basis for
  // refusing a write that overwrites somebody else's change, and two windows
  // pressing Accept both succeed with the second silently winning — the lost
  // update (P5 §A.8.1). Optional, so clients that do not send it are
  // unaffected; naturally idempotent, so a repeated claim by the same outlet
  // is still accepted rather than turned into a 412.
  const ifMatch = req.headers['if-match'];
  if (ifMatch && !matchesStrong(ifMatch, orderVersion(order))) {
    return sendProblem(
      res,
      preconditionFailed(
        'Order ini sudah ditangani oleh rekan kerja. Muat ulang untuk melihat keadaan terbaru.',
        req.originalUrl,
      ),
    );
  }

  const now = new Date().toISOString();
  if (order.outlet_id === null) {
    orderStore.assignOutlet(orderId, req.principal.outletId, now);
  }
  if (order.status !== 'processing') {
    orderStore.updateStatus(orderId, 'processing', now);
  }

  res.status(200).json(toOrderRepresentation(orderStore.getById(orderId)));
});

module.exports = router;